+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
+### Added
+- Added optional property `processedCardSelectionScenarioJsonString` to the following DTOs used by the non-Keyple KMP
+  client:
+  - `SelectAppAndAnalyzeContractsInputDto`
+  - `SelectAppAndLoadContractInputDto`
+  - `SelectAppAndPersonalizeCardInputDto`
 
 ## [2.0.2] - 2024-10-15
 ### Added